@@ -45,7 +45,9 @@ gem "prawnto_2",  :require => "prawnto"
 
 gem "prawn-table"
 
-gem "rails", ">= 7.0", "<7.1"
+gem "rack", "~> 2"
+
+gem "rails", ">= 7.1", "<7.2"
 
 gem "rails-i18n"
 
